fix(Authorization): Automatic silent refresh of tokens (on instantiation) is now controlled via configuration (disabled by default). (#245)
- The `AutorizationManager` will now only automatically attempt to
refresh tokens on creation when configured with `automaticSilentRefresh:
true`.
- The `startSilentRefresh` internal method has also been updated
dispatch an `authenticated` event after processing all refresh tokens.
- A new property `manager.waitForSilentRefresh` exposes a Promise
depicting the state of _all_ inflight refresh requests.

All of these changes are in service of better application state control
on instance creation. Without these changes, it was fairly common for
any application listening to changes to tokens to effectively bootstrap
twice - once for the token in storage and again when the tokens were
refreshed.