feat(Authorization): updates SDK methods to automatically attempt to refresh tokens when 401 errors are encountered. (#251)
feat(Authorization)!: `AuthorizationManager` no longer refreshes tokens
on initialization. If you would like to continue to use this behavior,
call `manager.refreshTokens()` directly from your application (after
initialization).
feat(Authorization): Events can be bound to an `AuthorizationManager` on
creation by passing an `events` property in the configuration. This
ensures listeners can be bound before initial (constructor) events
dispatch.
feat(Authorization): When a `manager` is provided to SDK methods, if a
`401` is encountered the method will attempt to refresh the previous
token used (when refresh tokens are enabled) and make the request again
with a new token. **This will only be attempted _once_, per method.**